rename pdf-selection-popup
